agent/requirements.txt
agent-framework[azure-ai]>=0.1.0
azure-identity>=1.15.0
azure-cosmos>=4.5.0
azure-search-documents>=11.4.0
python-dotenv>=1.0.0
asyncio
typing